+++ src/elements/card/src/stories.tsx
@@ -234,6 +234,67 @@ export const HeaderChildrenExample = () => {
   )
 }
 
+export const FeaturedExample = () => {
+  const title = text('Title', 'Title')
+  const linkText = text('Link text', 'Read more')
+  const superScript = text('Superscript', 'Superscript')
+
+  const content = text(
+    'Content',
+    'its a card with a picture of a really cute kitten'
+  )
+
+  const imageSize = select(
+    'Image Size',
+    { cover: 'cover', contain: 'contain' },
+    'cover'
+  )
+
+  const img =
+    'https://uswitch-contentful.imgix.net/t014ej9w3ur1/YPnGDSG9aTIPmg1rlWsZu/94483e7cec0dd6ac947e1f974650210f/800.jpg'
+
+  return (
+    <div>
+      <Container>
+        <Card
+          imgSrc={img}
+          imgAlt="Picture of a cute kitten"
+          imgSizes="(max-width: 33vw), 33vw"
+          imageSize={imageSize}
+          title={title}
+          description={content}
+          linkHref="https://www.uswitch.com"
+          variant="featured"
+          contentChildren={
+            <div>
+              <h2>Content elements</h2>
+              <span sx={{ color: 'red' }}>comes here</span>
+            </div>
+          }
+        />
+        <Card
+          imgSrc={img}
+          imgAlt="Picture of a cute kitten"
+          imgSizes="(max-width: 33vw), 33vw"
+          imageSize={imageSize}
+          title={title}
+          description={content}
+          linkHref="https://www.uswitch.com"
+          variant="featured"
+          linkText={linkText}
+          superScript={superScript}
+          contentChildren={
+            <div>
+              <h2>Content elements</h2>
+              <span sx={{ color: 'red' }}>comes here</span>
+            </div>
+          }
+        />
+      </Container>
+    </div>
+  )
+}
+
 HorizontalCards.story = {
   parameters: {
     percy: { skip: true }
@@ -252,12 +313,19 @@ HeaderChildrenExample.story = {
   }
 }
 
+FeaturedExample.story = {
+  parameters: {
+    percy: { skip: true }
+  }
+}
+
 export const AutomatedTests = () => {
   return (
     <AllThemes themes={['uswitch', 'money', 'broadband-deals']}>
       <VerticalCards />
       <HorizontalCards />
       <HeaderChildrenExample />
+      <FeaturedExample />
     </AllThemes>
   )
 }