+++ src/themes/save-on-energy/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.save-on-energy-theme@1.12.1...@uswitch/trustyle.save-on-energy-theme@1.12.2) (2020-08-19)
+
+**Note:** Version bump only for package @uswitch/trustyle.save-on-energy-theme
+
+
+
+
+
+## [1.12.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.save-on-energy-theme@1.12.0...@uswitch/trustyle.save-on-energy-theme@1.12.1) (2020-08-18)
+
+
+### Bug Fixes
+
+* add responsive card variant ([5799839](https://github.com/uswitch/trustyle/commit/5799839))
+
+
+
+
+
+
 # [1.12.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.save-on-energy-theme@1.11.2...@uswitch/trustyle.save-on-energy-theme@1.12.0) (2020-08-11)